Fix flaky appveyor tests
We were relying on the number of goroutines in those tests. This commit
replaces that logic with grepping through the current callstack.

The 'doesn't return after version negotiation is established if no
ConnState is defined' test is disabled, since I can't figure out a way
to make it work without leaking sessions. I filed #506 to track that
work.

Fixes #484.