make sure that the SHLO is always sent with secure encryption level
This is done as follows:
1. The packetPacker now first assembles the packet payload before
requesting the sealer from the crypto setup. This fixes a race condition
where the crypto setup could write the SHLO after the packetPacker
already fetched the unencrypted sealer.
2. The crypto setup keeps track of when the SHLO was written to the
crypto stream. Only after this has happened *and* after at the secure
sealer was used, it will allow usage of the forward-secure sealer. This
makes sure that the SHLO is never sent with forward-secure encryption.