fix integration tests on Chrome
The Chrome integration tests downloading and uploading 50 MB of data
fails on CircleCI and on Ubuntu 19.04. Chrome doesn't output any error
message, and also seems to completely calculate the array buffer
containing the random data, but it doesn't execute the AJAX requesting
to load the file from the QUIC server. Using different parameters for
the Lehmer random number generator apparently fixes the test.