quicvarint: add benchmark comparing 3 versions of Read
- Read(io.Reader) with type assertion to quicvarint.Reader
- Read(io.ByteReader) with no type assertion (current master branch)
- Read(quicvarint.Reader) with no type assertion

Results: the current Read func in this branch is approximately 86% slower
(1.86x) runtime than Read on current master branch. The other implementation
which accepts a quicvarint.Reader with no type assertion is on par with master.

Given this function is in the hot path, I recommend we eliminate the type
assertion in favor of the caller ensuring argument to Read() satisfies the
quicvarint.Reader or io.ByteReader interface.