http3: ContentLength must be -1 for unknown body length
According to the http.Request docs, on a server request, a ContentLength
of -1 indicates an unknown length.

So in the case of a request without an explicit Content-Length header,
the ContentLength was being set to 0, which explicitly indicates there's
no body to read. Whereas -1 would indicate to read until EOF.

When no body is expected, let ContentLength be 0.