http3: prefer IPv6 over IPv4 when dialing
I'm not sure this is the correct fix. It seems like we should be able to handle
multiple DNS responses, and fall back from IPv6 to IPv4 if IPv6 is not
available (for example). That would be a bigger change, and would probably need
to be done in the quic and not the http3 package.