qlog: defer creation of the qlog file in QLOGDIR until the first write
If the syscall to create the qlog file takes a non-negligible amount of
time, this can block the connection's run loop, or, even worse, the
server's run loop.

By creating the qlog file on the first write, it will be created by a
the qlogger's Go routine, and not be able to block the run loop. The
qlogger uses a buffered channel (capacity: 50). The qlogger will
therefore only block if more than 50 qlog events are queued before the
syscall returns.