remove conn from packetHandlerMap on early error
This fixes an issue where if we would return early here with an error we
would never clean up our reference in the packet handler map.

but, afaict these calls never error (outside of tests with mock) because
they don't actually do any IO. However, it's probably best to guard here
against a leaked connection that would happen if these calls ever did
return an error in the future.