Fix TestDatagramLoss failures due to a too eager context timeout
While datagram payloads are sent with intervals of 0.5ms (total 50ms),
the TLS handshake alone takes 1 RTT (10ms + crypto overhead). On top of
that, the send queue is currently limited to 32 packets and can block
while waiting for the buffer to clear (presumably by receiving ACKs).
Raise the previous timeout of 100ms to avoid early timeouts.

Improve the test to check for the exact number of expected DATAGRAM
frames by discarding too large packets that do not have a DATAGRAM
frame. We can do this when the test timeout is sufficiently high.

Fixes #4784