Replace RSA-1024 and Ed25519 test certs by P-256
In FIPS mode, signing with RSA keys of 1024 bits and use of Ed25519 are
not permitted, replace them by NIST P-256. This can also a little bit
faster which might matter for fuzzing, but it is really insignificant.
This fixes an explosion of errors while running:

    GODEBUG=fips140=on go test ./...

This also fixes failures with `import _ "crypto/tls/fipsonly"` present:

    GOEXPERIMENT=boringcrypto go test ./...