Add an accessor for the local EnabledDatagrams setting
Currently, there is no way to detect whether Datagrams are enabled
on an http3.Conn or http3.ClientConn after it has been created.
This means that any application that needs this information must pass
this flag in parallel to any component that requires it, and risks
bugs if the value is set incorrectly.

This change exposes the EnableDatagrams setting from each connection,
allowing applications to determine if Datagrams can be used by
inspecting the connection.