allow connection timer resets to earlier times
This can happen when a connection starts sending after being idle. In
the idle state, the timer is set for the idle timeout, e.g. 30s into the
future. Once we start sending packets again, we need to be able to set
the loss recovery timer to timestamps before that old idle timeout
value.